requirements.txt
streamlit>=1.39.0
httpx[http2]>=0.27.0
websockets>=12.0
python-dotenv>=1.0.1
